% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnorm.R
\name{gnorm}
\alias{gnorm}
\alias{dgnorm}
\alias{pgnorm}
\alias{qgnorm}
\alias{rgnorm}
\title{The generalized normal distribution}
\source{
\code{dgnorm}, \code{pgnorm}, \code{qgnorm} and\code{rgnorm} are all parametrized as in Version 1 of the following: https://en.wikipedia.org/wiki/Generalized_normal_distribution
Functions and documentation modeled after GammaDist functions
}
\usage{
\code{dgnorm(x, mu = 0, alpha = 1, beta = 1, log = FALSE)}
\code{pgnorm(q, mu = 0, alpha = 1, beta = 1, lower.tail = TRUE, log.p = FALSE)}
\code{qgnorm(p, mu = 0, alpha = 1, beta = 1, lower.tail = TRUE, log.p = FALSE)}
\code{rgnorm(n, mu = 0, alpha = 1, beta = 1)}
}
\arguments{
\item{\code{x}}{vector of quantiles}

\item{\code{p}}{vector of probabilities}

\item{\code{n}}{number of observations}

\item{\code{mu}}{location parameter}

\item{\code{alpha}}{scale parameter}

\item{\code{beta}}{shape parameter}
}
\description{
Density, distribution function and random generation for the generalized normal distribution
}
