% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnorm.R
\name{gnorm}
\alias{gnorm}
\alias{dgnorm}
\alias{pgnorm}
\alias{qgnorm}
\alias{rgnorm}
\title{The generalized normal distribution}
\source{
\code{dgnorm}, \code{pgnorm}, \code{qgnorm} and\code{rgnorm} are all parametrized as in Version 1 of the \href{https://en.wikipedia.org/wiki/Generalized_normal_distribution}{Generalized Normal Distribution Wikipedia page}.
Functions and documentation modeled after GammaDist functions
}
\usage{
\code{dgnorm(x, mu = 0, alpha = 1, beta = 1, log = FALSE)}
\code{pgnorm(q, mu = 0, alpha = 1, beta = 1, lower.tail = TRUE, log.p = FALSE)}
\code{qgnorm(p, mu = 0, alpha = 1, beta = 1, lower.tail = TRUE, log.p = FALSE)}
\code{rgnorm(n, mu = 0, alpha = 1, beta = 1)}
}
\arguments{
\item{\code{x}}{vector of quantiles}

\item{\code{p}}{vector of probabilities}

\item{\code{n}}{number of observations}

\item{\code{mu}}{location parameter}

\item{\code{alpha}}{scale parameter}

\item{\code{beta}}{shape parameter}
}
\description{
Density, distribution function and random generation for the generalized normal/exponential power distribution. \cr \cr
A generalized normal random variable \eqn{x} with parameters \eqn{\mu}, \eqn{\alpha > 0} and \eqn{\beta > 0} has density:\cr
\deqn{p(x) = \beta exp{-(|x - \mu|/\alpha)^\beta}/(2\alpha \Gamma(1/\beta)).} \cr
The mean and variance of \eqn{x} are \eqn{\mu} and \eqn{\alpha^2 \Gamma(3/\beta)/\Gamma(1/\beta)}, respectively.
}
